feat: 대본 입력 화면과 녹음 연습 화면 연결 | rename: 폴더 이름 수정
- Navigator를 이용하여 대본 입력 화면과 녹음 연습 화면을 연결함.
- scriptInput feature의 폴더 이름이 viewModel로 되어 있어서 controller로 변경함.

#TT-88 scriptInput
